Add getter and setter for the access token
Currently the access token is set in the constructor only, and when a Guzzle client is passed as a second parameter, the access token is being ignored.

Having a setter allows the user to change accounts dynamically and the getter allows us to test that the client has the right access token.